Hyunmin

+++ src/main/java/com/team5/projrental/chat/ChatMapper.java
@@ -29,7 +29,7 @@ public interface ChatMapper {
     //상대방의 firebaseToken값 필요. 나의 pic, iuser값 필요
     UserEntity selOtherPersonByLoginUser(ChatMsgInsDto dto);
 
-    UserEntity selChatUser(int iuser);
+    UserEntity selChatUser(Long iuser);
 
     int delBeforeChatIstatus(ChatMsgInsDto dto);
 }
