+++ src/main/java/com/team5/projrental/chat/model/ChatMsgSelDto.java
@@ -17,7 +17,7 @@ public class ChatMsgSelDto {
     private int ichat;
 
     @JsonIgnore
-    private int loginedIuser;
+    private Long loginedIuser;
 
     @JsonIgnore
     private int startIdx;
