+++ src/main/java/com/team5/projrental/payment/model/proc/GetDepositAndPriceFromProduct.java
@@ -10,7 +10,7 @@ import java.time.LocalDate;
 @AllArgsConstructor
 @NoArgsConstructor
 public class GetDepositAndPriceFromProduct {
-    private Integer iseller;
+    private Long iseller;
     private Integer rentalPrice;
     private Integer deposit;
     private Integer price;
